requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.31
